Update on "[quant] Finishing refactor for quantization test files"
Summary:
- we can put both fake quant module and observer module tests in the test_workflow_module.py
- added test_quantized_functional.py
- moved tests in test_numerics.py to test_quantize.py and removed test_numerics.py

Test Plan:
python test/test_quantization.py

Reviewers:
.

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
